+++ src/test/java/dk/kb/present/TestFiles.java
@@ -42,6 +42,7 @@ public class TestFiles {
     public static final String PVICA_DOMS_MIG_eaea0362 = "internal_test_files/domsMigrated/eaea0362-bbad-43ec-8d5e-07df6957923b-20240612T112310.xml";
     public static final String PVICA_DOMS_MIG_9779a1b2 = "internal_test_files/domsMigrated/9779a1b2-88ff-408c-9352-ad614615b2e7-20240531T022942.xml";
     public static final String PVICA_DOMS_MIG_9ed10d66 = "internal_test_files/domsMigrated/9ed10d66-4791-4390-8db9-bb0bed7aebcb-20240531T023244.xml";
+    public static final String PVICA_DOMS_MIG_82514cd9 = "internal_test_files/domsMigrated/82514cd9-c30c-4963-8b5c-7c7d32684c17.xml";
     public static final String PVICA_HOMEMADE_HOLDBACK_TEST_RECORD = "internal_test_files/homemade/badHoldbackFieldContents.xml";
 
     // From preservica 6, not in preservica 7 stage